fix: module 分割を最適化
