+++ src/services/markdown/extensions/messageBox.ts
@@ -1,4 +1,10 @@
-import { Lexer, Parser, Token, TokensList, MarkedExtension } from "marked";
+import {
+  Lexer,
+  Parser,
+  Token,
+  TokensList,
+  TokenizerAndRendererExtension,
+} from "marked";
 
 interface MessageBoxToken {
   type: string;
@@ -18,33 +24,29 @@ interface BlockTokenizerThis {
  * :::
  * という形式のマークダウンを赤背景のメッセージボックスにレンダリングする
  */
-export const messageBoxExtension: MarkedExtension = {
-  extensions: [
-    {
-      name: "message-box",
-      level: "block",
-      tokenizer(this: BlockTokenizerThis, src: string) {
-        const rule = /^:::message\s*\n([\s\S]*?)\n:::\s*(?:\n|$)/;
-        const match = rule.exec(src);
-        if (match) {
-          const token: MessageBoxToken = {
-            type: "message-box",
-            raw: match[0],
-            text: match[1].trim(),
-            tokens: this.lexer.inlineTokens(match[1].trim(), []),
-          };
-          return token;
-        }
-        return undefined;
-      },
-      renderer(this: { parser: Parser }, token: Token) {
-        // token が MessageBoxToken として扱えるかチェック
-        const messageToken = token as MessageBoxToken;
-        if (!messageToken || !messageToken.tokens) {
-          return "";
-        }
-        return `<div class="bg-red-200 rounded p-2">${this.parser.parseInline(messageToken.tokens)}</div>\n`;
-      },
-    },
-  ],
+export const messageBoxExtension: TokenizerAndRendererExtension = {
+  name: "message-box",
+  level: "block",
+  tokenizer(this: BlockTokenizerThis, src: string) {
+    const rule = /^:::message\s*\n([\s\S]*?)\n:::\s*(?:\n|$)/;
+    const match = rule.exec(src);
+    if (match) {
+      const token: MessageBoxToken = {
+        type: "message-box",
+        raw: match[0],
+        text: match[1].trim(),
+        tokens: this.lexer.inlineTokens(match[1].trim(), []),
+      };
+      return token;
+    }
+    return undefined;
+  },
+  renderer(this: { parser: Parser }, token: Token) {
+    // token が MessageBoxToken として扱えるかチェック
+    const messageToken = token as MessageBoxToken;
+    if (!messageToken || !messageToken.tokens) {
+      return "";
+    }
+    return `<div class="bg-red-200 rounded p-2">${this.parser.parseInline(messageToken.tokens)}</div>\n`;
+  },
 };
